requirements.txt
fastapi
uvicorn
httpx
pydantic
pandas
openpyxl
anthropic
google-genai
mcp
mcp[cli]
google.generativeai